src/requirements.txt
azure-storage-blob
cdsapi==0.6.1
Django==4.2.6
geocoder==1.38.1
geopandas==0.14.0
joblib==1.3.2
matplotlib==3.8.0
ngrok==0.12.0
numpy==1.26.1
ordereddict==1.1
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
scikit_learn==1.3.1
seaborn==0.13.0
setuptools==68.0.0
Shapely==2.0.2
streamlit==1.27.2
xarray==2023.9.0
netCDF4

